+++ eras/mary/impl/src/Cardano/Ledger/Mary/Value.hs
@@ -68,9 +68,9 @@ import Control.DeepSeq (NFData (..), deepseq, rwhnf)
 import Control.Exception (assert)
 import Control.Monad (forM_, guard, unless, when)
 import Control.Monad.ST (runST)
-import Data.Aeson (FromJSON, FromJSONKey, ToJSON (..), (.=))
+import Data.Aeson (FromJSON (..), FromJSONKey, ToJSON (..), (.:), (.=))
 import qualified Data.Aeson as Aeson
-import Data.Aeson.Types (ToJSONKey (..), toJSONKeyText)
+import Data.Aeson.Types (FromJSONKeyFunction (..), Parser, ToJSONKey (..), toJSONKeyText)
 import qualified Data.ByteString as BS
 import qualified Data.ByteString.Base16 as BS16
 import Data.ByteString.Short (ShortByteString)
@@ -98,7 +98,7 @@ import qualified Data.Semigroup as Semigroup (Sum (..))
 import Data.Set (Set)
 import qualified Data.Set as Set
 import Data.Text (Text)
-import Data.Text.Encoding (decodeLatin1)
+import Data.Text.Encoding (decodeLatin1, encodeUtf8)
 import Data.Word (Word16, Word32, Word64)
 import GHC.Generics (Generic)
 import NoThunks.Class (NoThunks (..), OnlyCheckWhnfNamed (..))
@@ -117,12 +117,6 @@ newtype AssetName = AssetName {assetNameBytes :: SBS.ShortByteString}
 instance Show AssetName where
   show = show . assetNameToBytesAsHex
 
-assetNameToBytesAsHex :: AssetName -> BS.ByteString
-assetNameToBytesAsHex = BS16.encode . SBS.fromShort . assetNameBytes
-
-assetNameToTextAsHex :: AssetName -> Text
-assetNameToTextAsHex = decodeLatin1 . assetNameToBytesAsHex
-
 instance DecCBOR AssetName where
   decCBOR = do
     an <- decCBOR
@@ -133,6 +127,30 @@ instance DecCBOR AssetName where
             assetNameToTextAsHex $
               AssetName an
       else pure $ AssetName an
+
+instance ToJSON AssetName where
+  toJSON = Aeson.String . assetNameToTextAsHex
+
+instance ToJSONKey AssetName where
+  toJSONKey = toJSONKeyText assetNameToTextAsHex
+
+instance FromJSON AssetName where
+  parseJSON = Aeson.withText "AssetName" assetNameFromHexParser
+
+instance FromJSONKey AssetName where
+  fromJSONKey = FromJSONKeyTextParser assetNameFromHexParser
+
+assetNameFromHexParser :: Text -> Parser AssetName
+assetNameFromHexParser t =
+  case BS16.decode (encodeUtf8 t) of
+    Left e -> fail $ "AssetName: invalid hex: " <> e
+    Right bs -> pure $ AssetName (SBS.toShort bs)
+
+assetNameToBytesAsHex :: AssetName -> BS.ByteString
+assetNameToBytesAsHex = BS16.encode . SBS.fromShort . assetNameBytes
+
+assetNameToTextAsHex :: AssetName -> Text
+assetNameToTextAsHex = decodeLatin1 . assetNameToBytesAsHex
 
 -- | Policy ID
 newtype PolicyID = PolicyID {policyID :: ScriptHash}
@@ -153,7 +171,7 @@ newtype PolicyID = PolicyID {policyID :: ScriptHash}
 
 -- | The MultiAssets map
 newtype MultiAsset = MultiAsset (Map PolicyID (Map AssetName Integer))
-  deriving (Show, Generic, ToJSON, EncCBOR)
+  deriving (Show, Generic, FromJSON, ToJSON, EncCBOR)
 
 instance Eq MultiAsset where
   MultiAsset x == MultiAsset y = pointWise (pointWise (==)) x y
@@ -389,11 +407,11 @@ instance ToKeyValuePairs MaryValue where
     , "policies" .= ps
     ]
 
-instance ToJSON AssetName where
-  toJSON = Aeson.String . assetNameToTextAsHex
-
-instance ToJSONKey AssetName where
-  toJSONKey = toJSONKeyText assetNameToTextAsHex
+instance FromJSON MaryValue where
+  parseJSON = Aeson.withObject "MaryValue" $ \o ->
+    MaryValue
+      <$> o .: "lovelace"
+      <*> o .: "policies"
 
 -- ========================================================================
 -- Compactible